move download into util.files
