+++ cibuildwheel/windows.py
@@ -27,7 +27,6 @@ from .util import (
     BuildSelector,
     call,
     combine_constraints,
-    download,
     find_compatible_wheel,
     find_uv,
     get_build_verbosity_extra_flags,
@@ -40,7 +39,7 @@ from .util import (
     unwrap,
     virtualenv,
 )
-from .util.files import extract_zip, move_file
+from .util.files import download, extract_zip, move_file
 
 
 def get_nuget_args(
